Add check for ADS1x15 connection (Issue #1)
Added a check for a return value of -1 from a read of the Lo_thresh register and print out a detailed message if it fails. Skip the other tests in this case (but keep looping.)

Also fixed a (benign) compiler warning regarding the values used for the pos_sat_val and neg_sat_val variables.